Great refactoring
- Fixed typo on exceptions folders
