+++ app/core/exceptions/auth.py
