+++ app/core/exceptions/generic.py
