+++ app/core/exceptions/note.py
